// src/app/models/order-models/orderproducts.ts
export interface IOrderProduct{
    customerOrderProductId: number;
    orderId: number;
    productId: number;
    categoryId: number;
    productName: string;
    productDescription: string;
    productBrand: string;
    productQuality: string;
    productPrice: number;
    productQuantity: number;
}